+++ packages/buttons/src/elements/ButtonGroup.tsx
@@ -12,7 +12,7 @@ import { useButtonGroup } from '@zendeskgarden/container-buttongroup';
 import { StyledButtonGroup } from '../styled';
 import { ButtonGroupContext } from '../utils/useButtonGroupContext';
 
-interface IButtonGroupProps extends HTMLAttributes<HTMLDivElement> {
+export interface IButtonGroupProps extends HTMLAttributes<HTMLDivElement> {
   selectedItem?: any;
   onSelect?: (item: any) => void;
 }