+++ packages/buttons/src/index.ts
@@ -6,10 +6,15 @@
  */
 
 export { default as Button } from './elements/Button';
+export type { IButtonProps, IIconProps } from './elements/Button';
 export { default as Anchor } from './elements/Anchor';
+export type { IAnchorProps } from './elements/Anchor';
 export { default as ButtonGroup } from './elements/ButtonGroup';
+export type { IButtonGroupProps } from './elements/ButtonGroup';
 export { default as ChevronButton } from './elements/ChevronButton';
 export { default as IconButton } from './elements/IconButton';
 export { default as SplitButton } from './elements/SplitButton';
 export { default as ToggleButton } from './elements/ToggleButton';
+export type { IToggleButtonProps } from './elements/ToggleButton';
 export { default as ToggleIconButton } from './elements/ToggleIconButton';
+export type { IToggleIconButtonProps } from './elements/ToggleIconButton';