+++ packages/modals/src/elements/Modal.spec.tsx
@@ -40,23 +40,27 @@ describe('Modal', () => {
   });
 
   describe('componentDidMount()', () => {
-    it('should disable overflow scrolling for body element', () => {
-      const { baseElement } = render(<BasicExample />);
+    it('should disable overflow scrolling for the html element', () => {
+      render(<BasicExample />);
 
-      expect(baseElement.style.overflow).toBe('hidden');
+      expect(document.querySelector('html')?.style.overflow).toBe('hidden');
     });
   });
 
   describe('componentWillUnmount()', () => {
-    it('should apply previous body overflow style', () => {
-      document.body.style.overflow = 'auto';
-      const { baseElement, unmount } = render(<BasicExample />);
+    it('should apply previous html overflow style', () => {
+      const htmlElement = document.querySelector('html');
 
-      expect(baseElement.style.overflow).toBe('hidden');
+      if (htmlElement) {
+        htmlElement.style.overflow = 'scroll';
 
-      unmount();
+        const { unmount } = render(<BasicExample />);
 
-      expect(baseElement.style.overflow).toBe('auto');
+        expect(htmlElement.style.overflow).toBe('hidden');
+        unmount();
+
+        expect(htmlElement.style.overflow).toBe('scroll');
+      }
     });
   });